+++ client/src/main/java/replicant/messages/AuthTokenMessage.java
@@ -2,6 +2,7 @@ package replicant.messages;
 
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
@@ -13,6 +14,7 @@ public final class AuthTokenMessage extends ClientToServerMessage {
     @JsOverlay
     public static final String TYPE = Messages.C2S_Type.AUTH;
 
+    @JsProperty(name = Messages.Auth.TOKEN)
     @Nullable
     private String token;