+++ client/src/main/java/replicant/messages/ClientToServerMessage.java
@@ -2,13 +2,16 @@ package replicant.messages;
 
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
+import replicant.shared.Messages;
 
 /**
  * The abstract message type that all messages that are sent to the server conform to.
  */
 @JsType(isNative = true, namespace = JsPackage.GLOBAL, name = "Object")
 public abstract class ClientToServerMessage extends AbstractMessage {
+    @JsProperty(name = Messages.Common.REQUEST_ID)
     int requestId;
 
     /**