+++ client/src/main/java/replicant/messages/CommandMessage.java
@@ -2,6 +2,7 @@ package replicant.messages;
 
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsProperty;
 import jsinterop.annotations.JsType;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
@@ -13,9 +14,11 @@ public final class CommandMessage extends ClientToServerMessage {
     @JsOverlay
     public static final String TYPE = Messages.C2S_Type.COMMAND;
 
+    @JsProperty(name = Messages.Command.NAME)
     @NonNull
     private String name;
 
+    @JsProperty(name = Messages.Command.PAYLOAD)
     @Nullable
     private Object payload;