+++ client/src/main/java/replicant/messages/Message.externs.js
@@ -1,0 +1,61 @@
+/** @externs */
+
+/** @record */
+function ReplicantMessage() {}
+
+/** @type {?} */
+ReplicantMessage.prototype.commandResult;
+
+/** @type {?} */
+ReplicantMessage.prototype.datasetAddress;
+
+/** @type {?} */
+ReplicantMessage.prototype.datasetAddresses;
+
+/** @type {?} */
+ReplicantMessage.prototype.datasetCacheVersion;
+
+/** @type {?} */
+ReplicantMessage.prototype.datasetCacheVersions;
+
+/** @type {?} */
+ReplicantMessage.prototype.entityChanges;
+
+/** @type {?} */
+ReplicantMessage.prototype.entityId;
+
+/** @type {?} */
+ReplicantMessage.prototype.entityTypeId;
+
+/** @type {?} */
+ReplicantMessage.prototype.filterParameter;
+
+/** @type {?} */
+ReplicantMessage.prototype.filterParameterSubscriptionChanges;
+
+/** @type {?} */
+ReplicantMessage.prototype.message;
+
+/** @type {?} */
+ReplicantMessage.prototype.name;
+
+/** @type {?} */
+ReplicantMessage.prototype.payload;
+
+/** @type {?} */
+ReplicantMessage.prototype.replicantSessionId;
+
+/** @type {?} */
+ReplicantMessage.prototype.requestId;
+
+/** @type {?} */
+ReplicantMessage.prototype.subscriptionChange;
+
+/** @type {?} */
+ReplicantMessage.prototype.subscriptionChanges;
+
+/** @type {?} */
+ReplicantMessage.prototype.token;
+
+/** @type {?} */
+ReplicantMessage.prototype.type;